begin writing query for issue(most complicated query)

+++ backlog4s-core/src/main/scala/com/github/chaabaj/backlog4s/datas/Project.scala
@@ -1,7 +1,7 @@
 package com.github.chaabaj.backlog4s.datas
 
 object ProjectT {
-  def id(value: Long): Id[Project] = Id(value)
+  def id(value: Int): Id[Project] = Id(value)
   def key(value: String): Key[Project] = Key(value)
 }
